backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
proxmoxer==2.0.1
requests==2.31.0
icalendar==5.0.12
feedparser==6.0.11
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-dateutil==2.9.0
